@@ -1,27 +1,28 @@
 {
-  "_from": "accepts@1.3.4",
-  "_id": "accepts@1.3.4",
+  "_from": "accepts@latest",
+  "_id": "accepts@1.3.5",
   "_inBundle": false,
-  "_integrity": "sha1-hiRnWMfdbSGmR0/whKR0DsBesh8=",
+  "_integrity": "sha1-63d99gEXI6OxTopywIBcjoZ0a9I=",
   "_location": "/accepts",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "tag",
     "registry": true,
-    "raw": "accepts@1.3.4",
+    "raw": "accepts@latest",
     "name": "accepts",
     "escapedName": "accepts",
-    "rawSpec": "1.3.4",
+    "rawSpec": "latest",
     "saveSpec": null,
-    "fetchSpec": "1.3.4"
+    "fetchSpec": "latest"
   },
   "_requiredBy": [
+    "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/accepts/-/accepts-1.3.4.tgz",
-  "_shasum": "86246758c7dd6d21a6474ff084a4740ec05eb21f",
-  "_spec": "accepts@1.3.4",
-  "_where": "/home/pluma/projects/arangodb/arangodb/js/node",
+  "_resolved": "https://registry.npmjs.org/accepts/-/accepts-1.3.5.tgz",
+  "_shasum": "eb777df6011723a3b14e8a72c0805c8e86746bd2",
+  "_spec": "accepts@latest",
+  "_where": "/mnt/c/Users/me/Documents/Projects/arangodb/arangodb/js/node",
   "bugs": {
     "url": "https://github.com/jshttp/accepts/issues"
   },
@@ -38,18 +39,18 @@
     }
   ],
   "dependencies": {
-    "mime-types": "~2.1.16",
+    "mime-types": "~2.1.18",
     "negotiator": "0.6.1"
   },
   "deprecated": false,
   "description": "Higher-level content negotiation",
   "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
+    "eslint": "4.18.1",
+    "eslint-config-standard": "11.0.0",
+    "eslint-plugin-import": "2.9.0",
     "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-node": "6.0.1",
+    "eslint-plugin-promise": "3.6.0",
     "eslint-plugin-standard": "3.0.1",
     "istanbul": "0.4.5",
     "mocha": "~1.21.5"
@@ -81,5 +82,5 @@
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
   },
-  "version": "1.3.4"
+  "version": "1.3.5"
 }
