@@ -1,17 +1,17 @@
 {
-  "_from": "ansi_up",
-  "_id": "ansi_up@2.0.2",
+  "_from": "ansi_up@latest",
+  "_id": "ansi_up@4.0.3",
   "_inBundle": false,
-  "_integrity": "sha1-m1TeUIxcV59baWjmXBuGPgaAq5I=",
+  "_integrity": "sha512-ub+xzfB7Mkikx6u0/iHx97Cvs5SJRNjh3yoQpuRaAvC9L2t8ZghlqG6jXIQ7XkbtQyHJU/EP2AnPvwcib0vu6A==",
   "_location": "/ansi_up",
   "_phantomChildren": {},
   "_requested": {
     "type": "tag",
     "registry": true,
-    "raw": "ansi_up",
+    "raw": "ansi_up@latest",
     "name": "ansi_up",
     "escapedName": "ansi_up",
-    "rawSpec": "",
+    "rawSpec": "latest",
     "saveSpec": null,
     "fetchSpec": "latest"
   },
@@ -19,10 +19,10 @@
     "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/ansi_up/-/ansi_up-2.0.2.tgz",
-  "_shasum": "9b54de508c5c579f5b6968e65c1b863e0680ab92",
-  "_spec": "ansi_up",
-  "_where": "/Users/manu/arangodb-devel/tmp",
+  "_resolved": "https://registry.npmjs.org/ansi_up/-/ansi_up-4.0.3.tgz",
+  "_shasum": "7cc6683304c71f4dcd83725b7e8d313089e369a7",
+  "_spec": "ansi_up@latest",
+  "_where": "/mnt/c/Users/me/Documents/Projects/arangodb/arangodb/js/node",
   "author": {
     "name": "drudru",
     "email": "drudru@gmail.com"
@@ -36,7 +36,7 @@
   "devDependencies": {
     "mocha": "*",
     "should": "*",
-    "typescript": "^2.1"
+    "typescript": "^3.0"
   },
   "engines": {
     "node": "*"
@@ -57,5 +57,6 @@
     "build": "make typescript",
     "test": "make test"
   },
-  "version": "2.0.2"
+  "types": "./ansi_up.d.ts",
+  "version": "4.0.3"
 }
