@@ -1,73 +1,69 @@
 {
-  "name": "dedent",
-  "version": "0.6.0",
+  "_from": "dedent@latest",
+  "_id": "dedent@0.7.0",
+  "_inBundle": false,
+  "_integrity": "sha1-JJXduvbrh0q7Dhvp3yLS5aVEMmw=",
+  "_location": "/dedent",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "tag",
+    "registry": true,
+    "raw": "dedent@latest",
+    "name": "dedent",
+    "escapedName": "dedent",
+    "rawSpec": "latest",
+    "saveSpec": null,
+    "fetchSpec": "latest"
+  },
+  "_requiredBy": [
+    "#USER",
+    "/"
+  ],
+  "_resolved": "https://registry.npmjs.org/dedent/-/dedent-0.7.0.tgz",
+  "_shasum": "2495ddbaf6eb874abb0e1be9df22d2e5a544326c",
+  "_spec": "dedent@latest",
+  "_where": "/mnt/c/Users/me/Documents/Projects/arangodb/arangodb/js/node",
+  "author": {
+    "name": "Desmond Brand",
+    "email": "dmnd@desmondbrand.com",
+    "url": "http://desmondbrand.com"
+  },
+  "bugs": {
+    "url": "https://github.com/dmnd/dedent/issues"
+  },
+  "bundleDependencies": false,
+  "deprecated": false,
   "description": "An ES6 string tag that strips indentation from multi-line strings",
-  "main": "dist/dedent.js",
+  "devDependencies": {
+    "babel-cli": "^6.22.2",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-es2016": "^6.22.0",
+    "babel-preset-es2017": "^6.22.0",
+    "eslint": "^3.14.1",
+    "jest": "^18.1.0"
+  },
   "files": [
     "dist/dedent.js",
     "LICENSE"
   ],
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/dmnd/dedent.git"
-  },
+  "homepage": "https://github.com/dmnd/dedent",
   "keywords": [
     "dedent",
     "tag",
     "multi-line string",
     "es6"
   ],
-  "author": {
-    "name": "Desmond Brand",
-    "email": "dmnd@desmondbrand.com",
-    "url": "http://desmondbrand.com"
-  },
   "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/dmnd/dedent/issues"
-  },
-  "homepage": "https://github.com/dmnd/dedent",
-  "devDependencies": {
-    "babel-jest": "^5.0.1",
-    "eslint": "^1.9.0",
-    "gulp": "^3.8.10",
-    "gulp-babel": "^5.1.0",
-    "jest-cli": "^0.7.1"
+  "main": "dist/dedent.js",
+  "name": "dedent",
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/dmnd/dedent.git"
   },
   "scripts": {
-    "test": "jest",
-    "lint": "eslint dedent.js __tests__"
+    "build": "babel dedent.js --out-file dist/dedent.js",
+    "lint": "eslint dedent.js __tests__",
+    "test": "jest"
   },
-  "jest": {
-    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
-    "testFileExtensions": [
-      "js"
-    ],
-    "moduleFileExtensions": [
-      "js",
-      "json"
-    ]
-  },
-  "gitHead": "e7d0bcea425ff647791a30cd9db0fb486c26078a",
-  "_id": "dedent@0.6.0",
-  "_shasum": "0e6da8f0ce52838ef5cec5c8f9396b0c1b64a3cb",
-  "_from": "dedent@latest",
-  "_npmVersion": "3.3.6",
-  "_nodeVersion": "5.0.0",
-  "_npmUser": {
-    "name": "dmnd",
-    "email": "dmnd@desmondbrand.com"
-  },
-  "maintainers": [
-    {
-      "name": "dmnd",
-      "email": "dmnd@desmondbrand.com"
-    }
-  ],
-  "dist": {
-    "shasum": "0e6da8f0ce52838ef5cec5c8f9396b0c1b64a3cb",
-    "tarball": "http://registry.npmjs.org/dedent/-/dedent-0.6.0.tgz"
-  },
-  "directories": {},
-  "_resolved": "https://registry.npmjs.org/dedent/-/dedent-0.6.0.tgz"
+  "version": "0.7.0"
 }
