@@ -1,68 +1,48 @@
 {
-  "_args": [
-    [
-      "ansi-escapes@^1.1.0",
-      "/home/pluma/projects/arango/arangodb/js/node/node_modules/eslint/node_modules/inquirer"
-    ]
-  ],
-  "_from": "ansi-escapes@>=1.1.0 <2.0.0",
-  "_id": "ansi-escapes@1.4.0",
-  "_inCache": true,
-  "_installable": true,
+  "_from": "ansi-escapes@^3.2.0",
+  "_id": "ansi-escapes@3.2.0",
+  "_inBundle": false,
+  "_integrity": "sha512-cBhpre4ma+U0T1oM5fXg7Dy1Jw7zzwv7lt/GoCpr+hDQJoYnKVPLL4dCvSEFMmQurOQvSrwT7SL/DAlhBI97RQ==",
   "_location": "/eslint/ansi-escapes",
-  "_nodeVersion": "4.4.2",
-  "_npmOperationalInternal": {
-    "host": "packages-16-east.internal.npmjs.com",
-    "tmp": "tmp/ansi-escapes-1.4.0.tgz_1460925437568_0.228597579523921"
-  },
-  "_npmUser": {
-    "email": "sindresorhus@gmail.com",
-    "name": "sindresorhus"
-  },
-  "_npmVersion": "2.15.0",
   "_phantomChildren": {},
   "_requested": {
+    "type": "range",
+    "registry": true,
+    "raw": "ansi-escapes@^3.2.0",
     "name": "ansi-escapes",
-    "raw": "ansi-escapes@^1.1.0",
-    "rawSpec": "^1.1.0",
-    "scope": null,
-    "spec": ">=1.1.0 <2.0.0",
-    "type": "range"
+    "escapedName": "ansi-escapes",
+    "rawSpec": "^3.2.0",
+    "saveSpec": null,
+    "fetchSpec": "^3.2.0"
   },
   "_requiredBy": [
     "/eslint/inquirer"
   ],
-  "_resolved": "https://registry.npmjs.org/ansi-escapes/-/ansi-escapes-1.4.0.tgz",
-  "_shasum": "d3a8a83b319aa67793662b13e761c7911422306e",
-  "_shrinkwrap": null,
-  "_spec": "ansi-escapes@^1.1.0",
-  "_where": "/home/pluma/projects/arango/arangodb/js/node/node_modules/eslint/node_modules/inquirer",
+  "_resolved": "https://registry.npmjs.org/ansi-escapes/-/ansi-escapes-3.2.0.tgz",
+  "_shasum": "8780b98ff9dbf5638152d1f1fe5c1d7b4442976b",
+  "_spec": "ansi-escapes@^3.2.0",
+  "_where": "/mnt/c/Users/me/Documents/Projects/arangodb/arangodb/js/node/node_modules/eslint/node_modules/inquirer",
   "author": {
-    "email": "sindresorhus@gmail.com",
     "name": "Sindre Sorhus",
+    "email": "sindresorhus@gmail.com",
     "url": "sindresorhus.com"
   },
   "bugs": {
     "url": "https://github.com/sindresorhus/ansi-escapes/issues"
   },
-  "dependencies": {},
+  "bundleDependencies": false,
+  "deprecated": false,
   "description": "ANSI escape codes for manipulating the terminal",
   "devDependencies": {
     "ava": "*",
     "xo": "*"
   },
-  "directories": {},
-  "dist": {
-    "shasum": "d3a8a83b319aa67793662b13e761c7911422306e",
-    "tarball": "https://registry.npmjs.org/ansi-escapes/-/ansi-escapes-1.4.0.tgz"
-  },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
     "index.js"
   ],
-  "gitHead": "763a11847148479dd315c2b9f81b001c94740415",
   "homepage": "https://github.com/sindresorhus/ansi-escapes#readme",
   "keywords": [
     "ansi",
@@ -90,15 +70,7 @@
     "iterm2"
   ],
   "license": "MIT",
-  "maintainers": [
-    {
-      "email": "sindresorhus@gmail.com",
-      "name": "sindresorhus"
-    }
-  ],
   "name": "ansi-escapes",
-  "optionalDependencies": {},
-  "readme": "ERROR: No README data found!",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/sindresorhus/ansi-escapes.git"
@@ -106,5 +78,5 @@
   "scripts": {
     "test": "xo && ava"
   },
-  "version": "1.4.0"
+  "version": "3.2.0"
 }
