@@ -1,66 +1,49 @@
 {
-  "_args": [
-    [
-      "ansi-regex@^2.0.0",
-      "/home/pluma/projects/arango/arangodb/js/node/node_modules/eslint/node_modules/has-ansi"
-    ]
-  ],
-  "_from": "ansi-regex@>=2.0.0 <3.0.0",
-  "_id": "ansi-regex@2.0.0",
-  "_inCache": true,
-  "_installable": true,
+  "_from": "ansi-regex@^3.0.0",
+  "_id": "ansi-regex@3.0.0",
+  "_inBundle": false,
+  "_integrity": "sha1-7QMXwyIGT3lGbAKWa922Bas32Zg=",
   "_location": "/eslint/ansi-regex",
-  "_nodeVersion": "0.12.5",
-  "_npmUser": {
-    "email": "sindresorhus@gmail.com",
-    "name": "sindresorhus"
-  },
-  "_npmVersion": "2.11.2",
   "_phantomChildren": {},
   "_requested": {
+    "type": "range",
+    "registry": true,
+    "raw": "ansi-regex@^3.0.0",
     "name": "ansi-regex",
-    "raw": "ansi-regex@^2.0.0",
-    "rawSpec": "^2.0.0",
-    "scope": null,
-    "spec": ">=2.0.0 <3.0.0",
-    "type": "range"
+    "escapedName": "ansi-regex",
+    "rawSpec": "^3.0.0",
+    "saveSpec": null,
+    "fetchSpec": "^3.0.0"
   },
   "_requiredBy": [
-    "/eslint/has-ansi",
-    "/eslint/inquirer",
     "/eslint/strip-ansi"
   ],
-  "_resolved": "https://registry.npmjs.org/ansi-regex/-/ansi-regex-2.0.0.tgz",
-  "_shasum": "c5061b6e0ef8a81775e50f5d66151bf6bf371107",
-  "_shrinkwrap": null,
-  "_spec": "ansi-regex@^2.0.0",
-  "_where": "/home/pluma/projects/arango/arangodb/js/node/node_modules/eslint/node_modules/has-ansi",
+  "_resolved": "https://registry.npmjs.org/ansi-regex/-/ansi-regex-3.0.0.tgz",
+  "_shasum": "ed0317c322064f79466c02966bddb605ab37d998",
+  "_spec": "ansi-regex@^3.0.0",
+  "_where": "/mnt/c/Users/me/Documents/Projects/arangodb/arangodb/js/node/node_modules/eslint/node_modules/strip-ansi",
   "author": {
-    "email": "sindresorhus@gmail.com",
     "name": "Sindre Sorhus",
+    "email": "sindresorhus@gmail.com",
     "url": "sindresorhus.com"
   },
   "bugs": {
-    "url": "https://github.com/sindresorhus/ansi-regex/issues"
+    "url": "https://github.com/chalk/ansi-regex/issues"
   },
-  "dependencies": {},
+  "bundleDependencies": false,
+  "deprecated": false,
   "description": "Regular expression for matching ANSI escape codes",
   "devDependencies": {
-    "mocha": "*"
-  },
-  "directories": {},
-  "dist": {
-    "shasum": "c5061b6e0ef8a81775e50f5d66151bf6bf371107",
-    "tarball": "https://registry.npmjs.org/ansi-regex/-/ansi-regex-2.0.0.tgz"
+    "ava": "*",
+    "xo": "*"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
     "index.js"
   ],
-  "gitHead": "57c3f2941a73079fa8b081e02a522e3d29913e2f",
-  "homepage": "https://github.com/sindresorhus/ansi-regex",
+  "homepage": "https://github.com/chalk/ansi-regex#readme",
   "keywords": [
     "ansi",
     "styles",
@@ -89,26 +72,14 @@
     "pattern"
   ],
   "license": "MIT",
-  "maintainers": [
-    {
-      "email": "sindresorhus@gmail.com",
-      "name": "sindresorhus"
-    },
-    {
-      "email": "jappelman@xebia.com",
-      "name": "jbnicolai"
-    }
-  ],
   "name": "ansi-regex",
-  "optionalDependencies": {},
-  "readme": "ERROR: No README data found!",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/sindresorhus/ansi-regex.git"
+    "url": "git+https://github.com/chalk/ansi-regex.git"
   },
   "scripts": {
-    "test": "mocha test/test.js",
-    "view-supported": "node test/viewCodes.js"
+    "test": "xo && ava",
+    "view-supported": "node fixtures/view-codes.js"
   },
-  "version": "2.0.0"
+  "version": "3.0.0"
 }
