@@ -1,67 +1,57 @@
 {
-  "_args": [
-    [
-      "ansi-styles@^2.2.1",
-      "/home/pluma/projects/arango/arangodb/js/node/node_modules/eslint/node_modules/chalk"
-    ]
-  ],
-  "_from": "ansi-styles@>=2.2.1 <3.0.0",
-  "_id": "ansi-styles@2.2.1",
-  "_inCache": true,
-  "_installable": true,
+  "_from": "ansi-styles@^3.2.1",
+  "_id": "ansi-styles@3.2.1",
+  "_inBundle": false,
+  "_integrity": "sha512-VT0ZI6kZRdTh8YyJw3SMbYm/u+NqfsAxEpWO0Pf9sq8/e94WxxOpPKx9FR1FlyCtOVDNOQ+8ntlqFxiRc+r5qA==",
   "_location": "/eslint/ansi-styles",
-  "_nodeVersion": "4.3.0",
-  "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/ansi-styles-2.2.1.tgz_1459197317833_0.9694824463222176"
-  },
-  "_npmUser": {
-    "email": "sindresorhus@gmail.com",
-    "name": "sindresorhus"
-  },
-  "_npmVersion": "3.8.3",
   "_phantomChildren": {},
   "_requested": {
+    "type": "range",
+    "registry": true,
+    "raw": "ansi-styles@^3.2.1",
     "name": "ansi-styles",
-    "raw": "ansi-styles@^2.2.1",
-    "rawSpec": "^2.2.1",
-    "scope": null,
-    "spec": ">=2.2.1 <3.0.0",
-    "type": "range"
+    "escapedName": "ansi-styles",
+    "rawSpec": "^3.2.1",
+    "saveSpec": null,
+    "fetchSpec": "^3.2.1"
   },
   "_requiredBy": [
-    "/eslint/chalk"
+    "/eslint/chalk",
+    "/eslint/slice-ansi"
   ],
-  "_resolved": "https://registry.npmjs.org/ansi-styles/-/ansi-styles-2.2.1.tgz",
-  "_shasum": "b432dd3358b634cf75e1e4664368240533c1ddbe",
-  "_shrinkwrap": null,
-  "_spec": "ansi-styles@^2.2.1",
-  "_where": "/home/pluma/projects/arango/arangodb/js/node/node_modules/eslint/node_modules/chalk",
+  "_resolved": "https://registry.npmjs.org/ansi-styles/-/ansi-styles-3.2.1.tgz",
+  "_shasum": "41fbb20243e50b12be0f04b8dedbf07520ce841d",
+  "_spec": "ansi-styles@^3.2.1",
+  "_where": "/mnt/c/Users/me/Documents/Projects/arangodb/arangodb/js/node/node_modules/eslint/node_modules/chalk",
   "author": {
-    "email": "sindresorhus@gmail.com",
     "name": "Sindre Sorhus",
+    "email": "sindresorhus@gmail.com",
     "url": "sindresorhus.com"
+  },
+  "ava": {
+    "require": "babel-polyfill"
   },
   "bugs": {
     "url": "https://github.com/chalk/ansi-styles/issues"
   },
-  "dependencies": {},
+  "bundleDependencies": false,
+  "dependencies": {
+    "color-convert": "^1.9.0"
+  },
+  "deprecated": false,
   "description": "ANSI escape codes for styling strings in the terminal",
   "devDependencies": {
-    "mocha": "*"
-  },
-  "directories": {},
-  "dist": {
-    "shasum": "b432dd3358b634cf75e1e4664368240533c1ddbe",
-    "tarball": "https://registry.npmjs.org/ansi-styles/-/ansi-styles-2.2.1.tgz"
+    "ava": "*",
+    "babel-polyfill": "^6.23.0",
+    "svg-term-cli": "^2.1.1",
+    "xo": "*"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
     "index.js"
   ],
-  "gitHead": "95c59b23be760108b6530ca1c89477c21b258032",
   "homepage": "https://github.com/chalk/ansi-styles#readme",
   "keywords": [
     "ansi",
@@ -86,21 +76,14 @@
     "text"
   ],
   "license": "MIT",
-  "maintainers": [
-    {
-      "email": "sindresorhus@gmail.com",
-      "name": "sindresorhus"
-    }
-  ],
   "name": "ansi-styles",
-  "optionalDependencies": {},
-  "readme": "ERROR: No README data found!",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/chalk/ansi-styles.git"
   },
   "scripts": {
-    "test": "mocha"
+    "screenshot": "svg-term --command='node screenshot' --out=screenshot.svg --padding=3 --width=55 --height=3 --at=1000 --no-cursor",
+    "test": "xo && ava"
   },
-  "version": "2.2.1"
+  "version": "3.2.1"
 }
