@@ -1,67 +1,56 @@
 {
-  "_args": [
-    [
-      "callsites@^0.2.0",
-      "/home/pluma/projects/arango/arangodb/js/node/node_modules/eslint/node_modules/caller-path"
-    ]
-  ],
-  "_from": "callsites@>=0.2.0 <0.3.0",
-  "_id": "callsites@0.2.0",
-  "_inCache": true,
-  "_installable": true,
+  "_from": "callsites@^3.0.0",
+  "_id": "callsites@3.1.0",
+  "_inBundle": false,
+  "_integrity": "sha512-P8BjAsXvZS+VIDUI11hHCQEv74YT67YUi5JJFNWIqL235sBmjX4+qx9Muvls5ivyNENctx46xQLQ3aTuE7ssaQ==",
   "_location": "/eslint/callsites",
-  "_npmUser": {
-    "email": "sindresorhus@gmail.com",
-    "name": "sindresorhus"
-  },
-  "_npmVersion": "1.4.6",
   "_phantomChildren": {},
   "_requested": {
+    "type": "range",
+    "registry": true,
+    "raw": "callsites@^3.0.0",
     "name": "callsites",
-    "raw": "callsites@^0.2.0",
-    "rawSpec": "^0.2.0",
-    "scope": null,
-    "spec": ">=0.2.0 <0.3.0",
-    "type": "range"
+    "escapedName": "callsites",
+    "rawSpec": "^3.0.0",
+    "saveSpec": null,
+    "fetchSpec": "^3.0.0"
   },
   "_requiredBy": [
-    "/eslint/caller-path"
+    "/eslint/parent-module"
   ],
-  "_resolved": "https://registry.npmjs.org/callsites/-/callsites-0.2.0.tgz",
-  "_shasum": "afab96262910a7f33c19a5775825c69f34e350ca",
-  "_shrinkwrap": null,
-  "_spec": "callsites@^0.2.0",
-  "_where": "/home/pluma/projects/arango/arangodb/js/node/node_modules/eslint/node_modules/caller-path",
+  "_resolved": "https://registry.npmjs.org/callsites/-/callsites-3.1.0.tgz",
+  "_shasum": "b3630abd8943432f54b3f0519238e33cd7df2f73",
+  "_spec": "callsites@^3.0.0",
+  "_where": "/mnt/c/Users/me/Documents/Projects/arangodb/arangodb/js/node/node_modules/eslint/node_modules/parent-module",
   "author": {
-    "email": "sindresorhus@gmail.com",
     "name": "Sindre Sorhus",
-    "url": "http://sindresorhus.com"
+    "email": "sindresorhus@gmail.com",
+    "url": "sindresorhus.com"
   },
   "bugs": {
     "url": "https://github.com/sindresorhus/callsites/issues"
   },
-  "dependencies": {},
+  "bundleDependencies": false,
+  "deprecated": false,
   "description": "Get callsites from the V8 stack trace API",
   "devDependencies": {
-    "mocha": "*"
-  },
-  "directories": {},
-  "dist": {
-    "shasum": "afab96262910a7f33c19a5775825c69f34e350ca",
-    "tarball": "https://registry.npmjs.org/callsites/-/callsites-0.2.0.tgz"
+    "ava": "^1.4.1",
+    "tsd": "^0.7.2",
+    "xo": "^0.24.0"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=6"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "index.d.ts"
   ],
-  "homepage": "https://github.com/sindresorhus/callsites",
+  "homepage": "https://github.com/sindresorhus/callsites#readme",
   "keywords": [
-    "callsites",
-    "callsite",
-    "v8",
     "stacktrace",
+    "v8",
+    "callsite",
+    "callsites",
     "stack",
     "trace",
     "function",
@@ -70,21 +59,13 @@
     "debug"
   ],
   "license": "MIT",
-  "maintainers": [
-    {
-      "email": "sindresorhus@gmail.com",
-      "name": "sindresorhus"
-    }
-  ],
   "name": "callsites",
-  "optionalDependencies": {},
-  "readme": "ERROR: No README data found!",
   "repository": {
     "type": "git",
-    "url": "git://github.com/sindresorhus/callsites.git"
+    "url": "git+https://github.com/sindresorhus/callsites.git"
   },
   "scripts": {
-    "test": "mocha"
+    "test": "xo && ava && tsd"
   },
-  "version": "0.2.0"
+  "version": "3.1.0"
 }
