@@ -1,77 +1,61 @@
 {
-  "_args": [
-    [
-      "chalk@^1.1.3",
-      "/home/pluma/projects/arango/arangodb/js/node/node_modules/eslint"
-    ]
-  ],
-  "_from": "chalk@>=1.1.3 <2.0.0",
-  "_id": "chalk@1.1.3",
-  "_inCache": true,
-  "_installable": true,
+  "_from": "chalk@^2.1.0",
+  "_id": "chalk@2.4.2",
+  "_inBundle": false,
+  "_integrity": "sha512-Mti+f9lpJNcwF4tWV8/OrTTtF1gZi+f8FqlyAdouralcFWFQWF2+NgCHShjkCb+IFBLq9buZwE1xckQU4peSuQ==",
   "_location": "/eslint/chalk",
-  "_nodeVersion": "0.10.32",
-  "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/chalk-1.1.3.tgz_1459210604109_0.3892582862172276"
-  },
-  "_npmUser": {
-    "email": "i.am.qix@gmail.com",
-    "name": "qix"
-  },
-  "_npmVersion": "2.14.2",
   "_phantomChildren": {},
   "_requested": {
+    "type": "range",
+    "registry": true,
+    "raw": "chalk@^2.1.0",
     "name": "chalk",
-    "raw": "chalk@^1.1.3",
-    "rawSpec": "^1.1.3",
-    "scope": null,
-    "spec": ">=1.1.3 <2.0.0",
-    "type": "range"
+    "escapedName": "chalk",
+    "rawSpec": "^2.1.0",
+    "saveSpec": null,
+    "fetchSpec": "^2.1.0"
   },
   "_requiredBy": [
     "/eslint",
-    "/eslint/inquirer",
-    "/eslint/table"
+    "/eslint/@babel/highlight",
+    "/eslint/inquirer"
   ],
-  "_resolved": "https://registry.npmjs.org/chalk/-/chalk-1.1.3.tgz",
-  "_shasum": "a8115c55e4a702fe4d150abd3872822a7e09fc98",
-  "_shrinkwrap": null,
-  "_spec": "chalk@^1.1.3",
-  "_where": "/home/pluma/projects/arango/arangodb/js/node/node_modules/eslint",
+  "_resolved": "https://registry.npmjs.org/chalk/-/chalk-2.4.2.tgz",
+  "_shasum": "cd42541677a54333cf541a49108c1432b44c9424",
+  "_spec": "chalk@^2.1.0",
+  "_where": "/mnt/c/Users/me/Documents/Projects/arangodb/arangodb/js/node/node_modules/eslint",
   "bugs": {
     "url": "https://github.com/chalk/chalk/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {
-    "ansi-styles": "^2.2.1",
-    "escape-string-regexp": "^1.0.2",
-    "has-ansi": "^2.0.0",
-    "strip-ansi": "^3.0.0",
-    "supports-color": "^2.0.0"
+    "ansi-styles": "^3.2.1",
+    "escape-string-regexp": "^1.0.5",
+    "supports-color": "^5.3.0"
   },
-  "description": "Terminal string styling done right. Much color.",
+  "deprecated": false,
+  "description": "Terminal string styling done right",
   "devDependencies": {
-    "coveralls": "^2.11.2",
-    "matcha": "^0.6.0",
-    "mocha": "*",
-    "nyc": "^3.0.0",
-    "require-uncached": "^1.0.2",
-    "resolve-from": "^1.0.0",
-    "semver": "^4.3.3",
+    "ava": "*",
+    "coveralls": "^3.0.0",
+    "execa": "^0.9.0",
+    "flow-bin": "^0.68.0",
+    "import-fresh": "^2.0.0",
+    "matcha": "^0.7.0",
+    "nyc": "^11.0.2",
+    "resolve-from": "^4.0.0",
+    "typescript": "^2.5.3",
     "xo": "*"
   },
-  "directories": {},
-  "dist": {
-    "shasum": "a8115c55e4a702fe4d150abd3872822a7e09fc98",
-    "tarball": "https://registry.npmjs.org/chalk/-/chalk-1.1.3.tgz"
-  },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "templates.js",
+    "types/index.d.ts",
+    "index.js.flow"
   ],
-  "gitHead": "0d8d8c204eb87a4038219131ad4d8369c9f59d24",
   "homepage": "https://github.com/chalk/chalk#readme",
   "keywords": [
     "color",
