@@ -1,66 +1,51 @@
 {
-  "_args": [
-    [
-      "cli-cursor@^1.0.1",
-      "/home/pluma/projects/arango/arangodb/js/node/node_modules/eslint/node_modules/inquirer"
-    ]
-  ],
-  "_from": "cli-cursor@>=1.0.1 <2.0.0",
-  "_id": "cli-cursor@1.0.2",
-  "_inCache": true,
-  "_installable": true,
+  "_from": "cli-cursor@^2.1.0",
+  "_id": "cli-cursor@2.1.0",
+  "_inBundle": false,
+  "_integrity": "sha1-s12sN2R5+sw+lHR9QdDQ9SOP/LU=",
   "_location": "/eslint/cli-cursor",
-  "_nodeVersion": "4.1.0",
-  "_npmUser": {
-    "email": "sindresorhus@gmail.com",
-    "name": "sindresorhus"
-  },
-  "_npmVersion": "2.14.3",
   "_phantomChildren": {},
   "_requested": {
+    "type": "range",
+    "registry": true,
+    "raw": "cli-cursor@^2.1.0",
     "name": "cli-cursor",
-    "raw": "cli-cursor@^1.0.1",
-    "rawSpec": "^1.0.1",
-    "scope": null,
-    "spec": ">=1.0.1 <2.0.0",
-    "type": "range"
+    "escapedName": "cli-cursor",
+    "rawSpec": "^2.1.0",
+    "saveSpec": null,
+    "fetchSpec": "^2.1.0"
   },
   "_requiredBy": [
     "/eslint/inquirer"
   ],
-  "_resolved": "https://registry.npmjs.org/cli-cursor/-/cli-cursor-1.0.2.tgz",
-  "_shasum": "64da3f7d56a54412e59794bd62dc35295e8f2987",
-  "_shrinkwrap": null,
-  "_spec": "cli-cursor@^1.0.1",
-  "_where": "/home/pluma/projects/arango/arangodb/js/node/node_modules/eslint/node_modules/inquirer",
+  "_resolved": "https://registry.npmjs.org/cli-cursor/-/cli-cursor-2.1.0.tgz",
+  "_shasum": "b35dac376479facc3e94747d41d0d0f5238ffcb5",
+  "_spec": "cli-cursor@^2.1.0",
+  "_where": "/mnt/c/Users/me/Documents/Projects/arangodb/arangodb/js/node/node_modules/eslint/node_modules/inquirer",
   "author": {
-    "email": "sindresorhus@gmail.com",
     "name": "Sindre Sorhus",
+    "email": "sindresorhus@gmail.com",
     "url": "sindresorhus.com"
   },
   "bugs": {
     "url": "https://github.com/sindresorhus/cli-cursor/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {
-    "restore-cursor": "^1.0.1"
+    "restore-cursor": "^2.0.0"
   },
+  "deprecated": false,
   "description": "Toggle the CLI cursor",
   "devDependencies": {
     "ava": "*",
     "xo": "*"
   },
-  "directories": {},
-  "dist": {
-    "shasum": "64da3f7d56a54412e59794bd62dc35295e8f2987",
-    "tarball": "https://registry.npmjs.org/cli-cursor/-/cli-cursor-1.0.2.tgz"
-  },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
     "index.js"
   ],
-  "gitHead": "6be5a384d90278c66aa30db5ecdec8dc68f17d4f",
   "homepage": "https://github.com/sindresorhus/cli-cursor#readme",
   "keywords": [
     "cli",
@@ -78,15 +63,7 @@
     "command-line"
   ],
   "license": "MIT",
-  "maintainers": [
-    {
-      "email": "sindresorhus@gmail.com",
-      "name": "sindresorhus"
-    }
-  ],
   "name": "cli-cursor",
-  "optionalDependencies": {},
-  "readme": "ERROR: No README data found!",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/sindresorhus/cli-cursor.git"
@@ -94,5 +71,8 @@
   "scripts": {
     "test": "xo && ava"
   },
-  "version": "1.0.2"
+  "version": "2.1.0",
+  "xo": {
+    "esnext": true
+  }
 }
