@@ -1266,9 +1266,13 @@ Result RocksDBCollection::insertDocument(arangodb::transaction::Methods* trx,
   RocksDBKeyLeaser key(trx);
   key->constructDocument(_objectId, documentId);
 
-  blackListKey(key->string().data(), static_cast<uint32_t>(key->string().size()));
-
-  RocksDBMethods* mthds = RocksDBTransactionState::toMethods(trx);
+  RocksDBTransactionState* state = RocksDBTransactionState::toState(trx);
+  if (state->hasHint(transaction::Hints::Hint::GLOBAL_MANAGED)) {
+    // blacklist new document to avoid caching without committing first
+    blackListKey(key.ref());
+  }
+    
+  RocksDBMethods* mthds = state->rocksdbMethods();
   // disable indexing in this transaction if we are allowed to
   IndexingDisabler disabler(mthds, trx->isSingleOperationTransaction());
 
@@ -1307,7 +1311,7 @@ Result RocksDBCollection::removeDocument(arangodb::transaction::Methods* trx,
   RocksDBKeyLeaser key(trx);
   key->constructDocument(_objectId, documentId);
 
-  blackListKey(key->string().data(), static_cast<uint32_t>(key->string().size()));
+  blackListKey(key.ref());
 
   RocksDBMethods* mthds = RocksDBTransactionState::toMethods(trx);
 
@@ -1349,14 +1353,15 @@ Result RocksDBCollection::updateDocument(transaction::Methods* trx,
   TRI_ASSERT(_objectId != 0);
   Result res;
 
-  RocksDBMethods* mthds = RocksDBTransactionState::toMethods(trx);
+  RocksDBTransactionState* state = RocksDBTransactionState::toState(trx);
+  RocksDBMethods* mthds = state->rocksdbMethods();
   // disable indexing in this transaction if we are allowed to
   IndexingDisabler disabler(mthds, trx->isSingleOperationTransaction());
 
   RocksDBKeyLeaser key(trx);
   key->constructDocument(_objectId, oldDocumentId);
   TRI_ASSERT(key->containsLocalDocumentId(oldDocumentId));
-  blackListKey(key->string().data(), static_cast<uint32_t>(key->string().size()));
+  blackListKey(key.ref());
 
   rocksdb::Status s = mthds->SingleDelete(RocksDBColumnFamily::documents(), key.ref());
   if (!s.ok()) {
@@ -1365,14 +1370,18 @@ Result RocksDBCollection::updateDocument(transaction::Methods* trx,
 
   key->constructDocument(_objectId, newDocumentId);
   TRI_ASSERT(key->containsLocalDocumentId(newDocumentId));
-  // simon: we do not need to blacklist the new documentId
   s = mthds->PutUntracked(RocksDBColumnFamily::documents(), key.ref(),
                           rocksdb::Slice(newDoc.startAs<char>(),
                                          static_cast<size_t>(newDoc.byteSize())));
   if (!s.ok()) {
     return res.reset(rocksutils::convertStatus(s, rocksutils::document));
   }
-
+  
+  if (state->hasHint(transaction::Hints::Hint::GLOBAL_MANAGED)) {
+    // blacklist new document to avoid caching without committing first
+    blackListKey(key.ref());
+  }
+    
   READ_LOCKER(guard, _indexesLock);
   for (std::shared_ptr<Index> const& idx : _indexes) {
     RocksDBIndex* rIdx = static_cast<RocksDBIndex*>(idx.get());
@@ -1754,12 +1763,13 @@ void RocksDBCollection::destroyCache() const {
 }
 
 // blacklist given key from transactional cache
-void RocksDBCollection::blackListKey(char const* data, std::size_t len) const {
+void RocksDBCollection::blackListKey(RocksDBKey const& k) const {
   if (useCache()) {
     TRI_ASSERT(_cache != nullptr);
     bool blacklisted = false;
     while (!blacklisted) {
-      auto status = _cache->blacklist(data, static_cast<uint32_t>(len));
+      auto status = _cache->blacklist(k.buffer()->data(),
+                                      static_cast<uint32_t>(k.buffer()->size()));
       if (status.ok()) {
         blacklisted = true;
       } else if (status.errorNumber() == TRI_ERROR_SHUTTING_DOWN) {
