@@ -77,7 +77,9 @@ class IRESEARCH_API scorers {
   /// @brief checks whether scorer with a specified name is registered
   ////////////////////////////////////////////////////////////////////////////////
   static bool exists(
-    const string_ref& name, const irs::text_format::type_id& args_format
+    const string_ref& name,
+    const irs::text_format::type_id& args_format,
+    bool load_library = true
   );
 
   ////////////////////////////////////////////////////////////////////////////////
@@ -89,7 +91,8 @@ class IRESEARCH_API scorers {
   static sort::ptr get(
     const string_ref& name,
     const irs::text_format::type_id& args_format,
-    const string_ref& args
+    const string_ref& args,
+    bool load_library = true
   ) NOEXCEPT;
 
   ////////////////////////////////////////////////////////////////////////////////
