@@ -739,6 +739,7 @@ static VPackBuilder assembleLocalCollectionInfo(
     }
     return ret;
   } catch (std::exception const& e) {
+    ret.clear();
     std::string errorMsg(
         "Maintenance::assembleLocalCollectionInfo: Failed to lookup database ");
     errorMsg += database;
@@ -789,6 +790,7 @@ static VPackBuilder assembleLocalDatabaseInfo(std::string const& database,
 
     return ret;
   } catch (std::exception const& e) {
+    ret.clear();
     std::string errorMsg(
         "Maintenance::assembleLocalDatabaseInfo: Failed to lookup database ");
     errorMsg += database;
@@ -819,7 +821,8 @@ arangodb::Result arangodb::maintenance::reportInCurrent(
 
     if (!cur.hasKey(cdbpath)) {
       auto const localDatabaseInfo = assembleLocalDatabaseInfo(dbName, allErrors);
-      if (!localDatabaseInfo.slice().isEmptyObject()) {
+      TRI_ASSERT(!localDatabaseInfo.slice().isNone());
+      if (!localDatabaseInfo.slice().isEmptyObject() && !localDatabaseInfo.slice().isNone()) {
         report.add(VPackValue(CURRENT_DATABASES + dbName + "/" + serverId));
         {
           VPackObjectBuilder o(&report);
@@ -841,7 +844,8 @@ arangodb::Result arangodb::maintenance::reportInCurrent(
             assembleLocalCollectionInfo(shSlice, shardMap.slice().get(shName),
                                         dbName, shName, serverId, allErrors);
         // Collection no longer exists
-        if (localCollectionInfo.slice().isEmptyObject()) {
+        TRI_ASSERT(!localCollectionInfo.slice().isNone());
+        if (localCollectionInfo.slice().isEmptyObject() || localCollectionInfo.slice().isNone()) {
           continue;
         }
 
