@@ -124,17 +124,16 @@ class KShortestPathsFinder : public ShortestPathFinder {
     VertexRef _pred;
     Edge _edge;
     double _weight;
+    bool _done;
 
-    // Using negative weight to signifiy start/end vertex
     explicit FoundVertex(VertexRef const& vertex)
-        : _vertex(vertex), _weight(0) {}
+        : _vertex(vertex), _weight(0), _done(false) {}
     FoundVertex(VertexRef const& vertex, VertexRef const& pred, Edge&& edge, double weight)
-        : _vertex(vertex), _pred(pred), _edge(std::move(edge)), _weight(weight) {}
+        : _vertex(vertex), _pred(pred), _edge(std::move(edge)), _weight(weight), _done(false) {}
     double weight() const { return _weight; }
     void setWeight(double weight) { _weight = weight; }
     VertexRef const& getKey() const { return _vertex; }
   };
-  //  typedef std::deque<VertexRef> Frontier;
   typedef ShortestPathPriorityQueue<VertexRef, FoundVertex, double> Frontier;
 
   // Contains the vertices that were found while searching
@@ -152,6 +151,7 @@ class KShortestPathsFinder : public ShortestPathFinder {
     Ball(VertexRef const& centre, Direction direction)
         : _centre(centre), _direction(direction) {
       auto v = std::make_unique<FoundVertex>(centre);
+      v->_done = true;
       _frontier.insert(centre, std::move(v));
     }
     ~Ball() {
