@@ -39,11 +39,14 @@
 #include "Aql/Query.h"
 #include "Aql/QueryString.h"
 #include "Basics/StaticStrings.h"
+#include "Basics/StringUtils.h"
+#include "Cluster/ClusterComm.h"
 #include "Cluster/ClusterInfo.h"
 #include "Cluster/ServerState.h"
 #include "IResearchAnalyzerFeature.h"
 #include "IResearchCommon.h"
 #include "Logger/LogMacros.h"
+#include "RestHandler/RestVocbaseBaseHandler.h"
 #include "RestServer/DatabaseFeature.h"
 #include "RestServer/QueryRegistryFeature.h"
 #include "RestServer/SystemDatabaseFeature.h"
@@ -277,6 +280,47 @@ bool equalAnalyzer(
          && features == pool.features(); // same features
 }
 
+////////////////////////////////////////////////////////////////////////////////
+/// @brief get the collection containing analyzer definitions (if found)
+///        taken from vocabse for single-server and from ClusterInfo on cluster
+/// @note cannot use arangodb::methods::Collections::lookup(...) since it will
+///       try to resolve via vocbase for the case of db-server
+/// @note cannot use arangodb::CollectionNameResolver::getCollection(...) since
+///       it will try to resolve via vocbase for the case of db-server
+////////////////////////////////////////////////////////////////////////////////
+std::shared_ptr<arangodb::LogicalCollection> getAnalyzerCollection( // get collection
+    TRI_vocbase_t const& vocbase // collection vocbase
+) {
+  if (arangodb::ServerState::instance()->isSingleServer()) {
+    return vocbase.lookupCollection(ANALYZER_COLLECTION_NAME);
+  }
+
+  try {
+    auto* ci = arangodb::ClusterInfo::instance();
+
+    if (ci) {
+      return ci->getCollectionNT(vocbase.name(), ANALYZER_COLLECTION_NAME);
+    }
+
+    LOG_TOPIC(, WARN, arangodb::iresearch::TOPIC)
+      << "failure to find 'ClusterInfo' instance while looking up Analyzer collection '" << ANALYZER_COLLECTION_NAME << "' in vocbase '" << vocbase.name() << "'";
+  } catch (arangodb::basics::Exception& e) {
+    LOG_TOPIC(, WARN, arangodb::iresearch::TOPIC)
+      << "caught exception while looking up Analyzer collection '" << ANALYZER_COLLECTION_NAME << "' in vocbase '" << vocbase.name() << "': " << e.code() << " " << e.what();
+    IR_LOG_EXCEPTION();
+  } catch (std::exception& e) {
+    LOG_TOPIC(, WARN, arangodb::iresearch::TOPIC)
+      << "caught exception while looking up Analyzer collection '" << ANALYZER_COLLECTION_NAME << "' in vocbase '" << vocbase.name() << "': " << e.what();
+    IR_LOG_EXCEPTION();
+  } catch (...) {
+    LOG_TOPIC(, WARN, arangodb::iresearch::TOPIC)
+      << "caught exception while looking up Analyzer collection '" << ANALYZER_COLLECTION_NAME << "' in vocbase '" << vocbase.name() << "'";
+    IR_LOG_EXCEPTION();
+  }
+
+  return nullptr;
+}
+
 std::string normalizedAnalyzerName(
     std::string database, // database
     irs::string_ref const& analyzer // analyzer
@@ -462,22 +506,11 @@ arangodb::Result visitAnalyzers( // visit analyzers
   TRI_vocbase_t& vocbase, // vocbase to visit
   std::function<arangodb::Result(arangodb::velocypack::Slice const& slice)> const& visitor // visitor
 ) {
-  if (arangodb::ServerState::instance()->isClusterRole()) {
-    static const auto queryString = arangodb::aql::QueryString( // query to execute
-      std::string("FOR d IN ") + ANALYZER_COLLECTION_NAME + " RETURN d" // query
-    );
-    arangodb::aql::Query query( // query
-      false, vocbase, queryString, nullptr, nullptr, arangodb::aql::PART_MAIN // args
-    );
-    auto* queryRegistry = arangodb::QueryRegistryFeature::registry();
-    auto result = query.executeSync(queryRegistry);
-
-    if (result.result.fail()) {
-      return result.result;
-    }
-
-    auto slice = result.data->slice();
-
+  static const auto resultVisitor = [](
+    std::function<arangodb::Result(arangodb::velocypack::Slice const& slice)> const& visitor, // visitor
+    TRI_vocbase_t const& vocbase, // vocbase
+    arangodb::velocypack::Slice const& slice // slice to visit
+  )->arangodb::Result {
     if (!slice.isArray()) {
       return arangodb::Result( // result
         TRI_ERROR_INTERNAL, // code
@@ -494,6 +527,117 @@ arangodb::Result visitAnalyzers( // visit analyzers
     }
 
     return arangodb::Result();
+  };
+
+  // FIXME TODO find a better way to query a cluster collection
+  // workaround for aql::Query failing to execute on a cluster collection
+  if (arangodb::ServerState::instance()->isDBServer()) {
+    auto cc = arangodb::ClusterComm::instance();
+
+    if (!cc) {
+      return arangodb::Result( // result
+        TRI_ERROR_INTERNAL, // code
+        std::string("failure to find 'ClusterComm' instance while visiting Analyzer collection '") + ANALYZER_COLLECTION_NAME + "' in vocbase '" + vocbase.name() + "'"
+      );
+    }
+
+    auto collection = getAnalyzerCollection(vocbase);
+
+    if (!collection) {
+      return arangodb::Result(); // nothing to load
+    }
+
+    static const std::string body("{}"); // RestSimpleQueryHandler::allDocuments() expects opbject (calls get() on slice)
+    std::vector<arangodb::ClusterCommRequest> requests;
+
+    // create a request for every shard
+    //for (auto& entry: collection->errorNum()) {
+    for (auto& entry: *(collection->shardIds())) {
+      auto& shardId = entry.first;
+      auto url = // url
+        "/_db/" + arangodb::basics::StringUtils::urlEncode(vocbase.name())
+        + arangodb::RestVocbaseBaseHandler::SIMPLE_QUERY_ALL_PATH
+        + "?collection=" + shardId;
+
+      requests.emplace_back( // add shard request
+        "shard:" + shardId, // shard
+        arangodb::rest::RequestType::PUT, // request type as per SimpleQueryHandker
+        url, // request url
+        std::shared_ptr<std::string const>(&body, [](std::string const*)->void {}) // body
+      );
+    }
+
+    // same timeout as in ClusterMethods::getDocumentOnCoordinator()
+    cc->performRequests( // execute requests
+      requests, 120.0, arangodb::iresearch::TOPIC, false, false // args
+    );
+
+    for (auto& request: requests) {
+      if (TRI_ERROR_ARANGO_DATA_SOURCE_NOT_FOUND == request.result.errorCode) {
+        continue; // treat missing collection as if there are no analyzers
+      }
+
+      if (TRI_ERROR_NO_ERROR != request.result.errorCode) {
+        return arangodb::Result( // result
+          request.result.errorCode, request.result.errorMessage // args
+        );
+      }
+
+      if (!request.result.answer) {
+        return arangodb::Result( // result
+          TRI_ERROR_INTERNAL, // code
+          std::string("failed to get answer from 'ClusterComm' instance while visiting Analyzer collection '") + ANALYZER_COLLECTION_NAME + "' in vocbase '" + vocbase.name() + "'"
+        );
+      }
+
+      auto slice = request.result.answer->payload();
+
+      if (!slice.hasKey("result")) {
+        return arangodb::Result( // result
+          TRI_ERROR_INTERNAL, // code
+          std::string("failed to parse result from 'ClusterComm' instance while visiting Analyzer collection '") + ANALYZER_COLLECTION_NAME + "' in vocbase '" + vocbase.name() + "'"
+        );
+      }
+
+      auto res = resultVisitor(visitor, vocbase, slice.get("result"));
+
+      if (!res.ok()) {
+        return res;
+      }
+    }
+
+    return arangodb::Result();
+  }
+
+  if (arangodb::ServerState::instance()->isClusterRole()) {
+    if (!getAnalyzerCollection(vocbase)) {
+      return arangodb::Result(); // treat missing collection as if there are no analyzers
+    }
+
+    static const auto queryString = arangodb::aql::QueryString( // query to execute
+      std::string("FOR d IN ") + ANALYZER_COLLECTION_NAME + " RETURN d" // query
+    );
+    arangodb::aql::Query query( // query
+      false, vocbase, queryString, nullptr, nullptr, arangodb::aql::PART_MAIN // args
+    );
+    auto* queryRegistry = arangodb::QueryRegistryFeature::registry();
+    auto result = query.executeSync(queryRegistry);
+
+    if (TRI_ERROR_ARANGO_DATA_SOURCE_NOT_FOUND == result.result.errorNumber()) {
+      return arangodb::Result(); // treat missing collection as if there are no analyzers
+    }
+
+    if (result.result.fail()) {
+      return result.result;
+    }
+
+    auto slice = result.data->slice();
+
+    return resultVisitor(visitor, vocbase, slice);
+  }
+
+  if (!vocbase.lookupCollection(ANALYZER_COLLECTION_NAME)) {
+    return arangodb::Result(); // treat missing collection as if there are no analyzers
   }
 
   arangodb::OperationOptions options;
@@ -517,22 +661,7 @@ arangodb::Result visitAnalyzers( // visit analyzers
 
   auto slice = arangodb::velocypack::Slice(result.buffer->data());
 
-  if (!slice.isArray()) {
-    return arangodb::Result( // result
-      TRI_ERROR_INTERNAL, // code
-      std::string("failed to parse contents of collection '") + ANALYZER_COLLECTION_NAME + "' in database '" + vocbase.name() + " while visiting analyzers"
-    );
-  }
-
-  for (arangodb::velocypack::ArrayIterator itr(slice); itr.valid(); ++itr) {
-    auto res = visitor(itr.value().resolveExternal());
-
-    if (!res.ok()) {
-      return res;
-    }
-  }
-
-  return arangodb::Result();
+  return resultVisitor(visitor, vocbase, slice);
 }
 
 typedef irs::async_utils::read_write_mutex::read_mutex ReadMutex;
@@ -553,7 +682,9 @@ IResearchAnalyzerFeature::AnalyzerPool::Builder::make(irs::string_ref const& typ
 
   // ArangoDB, for API consistency, only supports analyzers configurable via
   // jSON
-  return irs::analysis::analyzers::get(type, irs::text_format::json, properties);
+  return irs::analysis::analyzers::get( // get analyzer
+    type, irs::text_format::json, properties, false // args
+  );
 }
 
 IResearchAnalyzerFeature::AnalyzerPool::AnalyzerPool(irs::string_ref const& name)
@@ -821,15 +952,29 @@ arangodb::Result IResearchAnalyzerFeature::ensure( // ensure analyzer existence 
   bool allowCreation
 ) {
   try {
-    WriteMutex mutex(_mutex);
-    SCOPED_LOCK(mutex);
     auto split = splitAnalyzerName(name);
 
-    if (!split.first.null()) { // do not trigger load for static-analyzer requests
-      auto res = loadAnalyzers(split.first);
+    WriteMutex mutex(_mutex);
+    SCOPED_LOCK(mutex);
 
-      if (!res.ok()) {
-        return res;
+    if (!split.first.null()) { // do not trigger load for static-analyzer requests
+      // do not trigger load of analyzers on db-server to avoid recursive lock
+      // aquisition in ClusterInfo::loadPlan() if called due IResearchLink creation,
+      // also avoids extra cluster calls if it can be helped (optimization)
+      if (allowCreation && arangodb::ServerState::instance()->isDBServer()) {
+        auto itr = _analyzers.find( // find analyzer previous definition
+         irs::make_hashed_ref(name, std::hash<irs::string_ref>())
+        );
+
+        if (itr != _analyzers.end()) {
+          _analyzers.erase(itr); // remove old definition instead of reloading all
+        }
+      } else { // trigger analyzer load
+        auto res = loadAnalyzers(split.first);
+
+        if (!res.ok()) {
+          return res;
+        }
       }
     }
 
@@ -1132,6 +1277,10 @@ arangodb::Result IResearchAnalyzerFeature::loadAnalyzers( // load
 
     if (!engine || engine->inRecovery()) {
       // always load if inRecovery since collection contents might have changed
+      // unless on db-server which does not store analyzer definitions in collections
+      if (arangodb::ServerState::instance()->isDBServer()) {
+        return arangodb::Result(); // db-server should not access cluster during inRecovery
+      }
     } else if (arangodb::ServerState::instance()->isSingleServer()) { // single server
       if(itr != _lastLoad.end()) {
         return arangodb::Result(); // do not reload on single-server
@@ -1177,24 +1326,13 @@ arangodb::Result IResearchAnalyzerFeature::loadAnalyzers( // load
         TRI_ERROR_ARANGO_DATABASE_NOT_FOUND, // code
         std::string("failed to find database '") + std::string(database) + " while loading analyzers"
       );
-    } else {
-      std::shared_ptr<arangodb::LogicalCollection> collection;
-      auto collectionCallback = [&collection]( // store collection
-        std::shared_ptr<arangodb::LogicalCollection> const& col // args
-      )->void {
-        collection = col;
-      };
+    }
 
-      arangodb::methods::Collections::lookup( // find collection
-        *vocbase, ANALYZER_COLLECTION_NAME, collectionCallback // args
-      );
+    if (!getAnalyzerCollection(*vocbase)) {
+      cleanupAnalyzers(*this, itr, database); // cleanup any analyzers for 'database'
+      _lastLoad[databaseKey] = currentTimestamp; // update timestamp
 
-      if (!collection) {
-        cleanupAnalyzers(*this, itr, database); // cleanup any analyzers for 'database'
-        _lastLoad[databaseKey] = currentTimestamp; // update timestamp
-
-        return arangodb::Result(); // no collection means nothing to load
-      }
+      return arangodb::Result(); // no collection means nothing to load
     }
 
     Analyzers analyzers;
@@ -1681,18 +1819,14 @@ arangodb::Result IResearchAnalyzerFeature::storeAnalyzer(AnalyzerPool& pool) {
   }
 
   try {
-    std::shared_ptr<arangodb::LogicalCollection> collection;
-    auto collectionCallback = [&collection]( // store collection
-      std::shared_ptr<arangodb::LogicalCollection> const& col // args
-    )->void {
-      collection = col;
-    };
-
-    arangodb::methods::Collections::lookup( // find collection
-      *vocbase, ANALYZER_COLLECTION_NAME, collectionCallback // args
-    );
+    auto collection = getAnalyzerCollection(*vocbase);
 
     if (!collection) {
+      auto collectionCallback = [&collection]( // store collection
+        std::shared_ptr<arangodb::LogicalCollection> const& col // args
+      )->void {
+        collection = col;
+      };
       static auto const properties = // analyzer collection properties
         arangodb::velocypack::Parser::fromJson("{ \"isSystem\": true }");
       auto res = arangodb::methods::Collections::create( // create collection
